ppx_deriving.map2
Similar to map, but takes a pair as an argument.

Non-polymorphic arguments are copied from the 1st element of the
pair.

With variants, constructors must match in both arguments -
similarly with options and list and array lengths (otherwise
Invalid_argument is raised).  That is the shape of the agruments
must be the same.

Ptyp_alias is not implemented as I dont really understand this
part of the type system.